feat(invites): add canonical invite landing page and update routing for invite links
